requirements.txt
fastapi>=0.104.0
fastmcp>=0.2.0
httpx>=0.26.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
uvicorn>=0.24.0
uvloop>=0.19.0
jinja2>=3.1.0
websockets>=11.0
# Dependencias para PostgreSQL y búsqueda híbrida
asyncpg>=0.29.0
openai>=1.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
# Dependencias para el panel de administración
bcrypt>=4.0.0
PyJWT>=2.8.0
aiofiles>=23.0.0
python-multipart>=0.0.6
# Dependencias para el agente multi-agent
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
langgraph>=0.0.26
langchain-community>=0.0.10
langchain-mcp>=0.1.0
langchain-mcp-adapters>=0.1.0 